+++ app/Http/Controllers/ContactFormSubmissionController.php
@@ -2,13 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\ContactFormSubmissionRequest;
 use App\Models\ContactFormSubmission;
 use Illuminate\Http\Request;
 
 class ContactFormSubmissionController extends Controller
 {
-    
     public function __construct()
     {
         $this->middleware('auth');
@@ -19,20 +17,5 @@ class ContactFormSubmissionController extends Controller
         $submissions = ContactFormSubmission::paginate(4);
 
         return view('pages.admin.contact-form-submissions', compact('submissions'));
-    }
-
-    public function processContactFormSubmission(ContactFormSubmissionRequest $request)
-    {
-        // validate
-        $data = $request->validated();
-
-        // save data
-        ContactFormSubmission::create($data);
-
-        // TODO: send response email
-
-        // redirect with success message
-        return redirect()->route('site.contact')
-        ->with('success', 'Your contact request submitted successfully.');
     }
 }
